# backend/app/Http/Controllers/Api/GameController.php
<?php

namespace App\Http\Controllers\Api;

use App\Http\Controllers\Controller;
use App\Models\GameSession;
use App\Models\Question;
use App\Models\GameMatch;
use App\Models\Category;
use App\Services\GameService;
use Illuminate\Http\Request;
use OpenApi\Attributes as OA;
use Illuminate\Support\Facades\Gate;

class GameController extends Controller
{
    private $gameService;

    public function __construct(GameService $gameService)
    {
        $this->gameService = $gameService;
    }

    #[OA\Post(path: "/games", summary: "Start a new game session", tags: ["Game"])]
    #[OA\Response(response: 200, description: "Returns the newly created game session and first question")]
    public function store(Request $request)
    {
        // Resolve categories: support both database IDs (integers/numeric) and slugs (strings)
        $categoryInputs = $request->input('categories', []);
        $categoryIds = [];
        if (!empty($categoryInputs)) {
            foreach ($categoryInputs as $cat) {
                if (is_numeric($cat)) {
                    $categoryIds[] = (int) $cat;
                } elseif (is_string($cat)) {
                    $id = Category::where('slug', $cat)->value('id');
                    if ($id) $categoryIds[] = $id;
                }
            }
            $categoryIds = array_unique($categoryIds);
        }

        $session = $this->gameService->createSession(
            $request->user(),
            $request->input('match_id'),
            $categoryIds ?: null   // null = no filter (all categories)
        );

        $opponent = null;
        $opponents = [];
        if ($session->match_id) {
            $match = GameMatch::find($session->match_id);
            if ($match && $match->mode === '1v1') {
                $opponent = collect($match->players)->firstWhere('id', '!=', $request->user()->id);
            } elseif ($match && $match->mode === 'battle') {
                $opponents = collect($match->players)->filter(function($p) use ($request) {
                    return $p['id'] !== $request->user()->id;
                })->values()->all();
            }
        }

        return response()->json([
            'session'  => $session,
            'question' => $this->gameService->getNextQuestion($session),
            'opponent' => $opponent,
            'opponents' => $opponents
        ]);
    }


    #[OA\Get(path: "/games/{session}", summary: "Get an active game session state", tags: ["Game"])]
    #[OA\Parameter(name: "session", in: "path", required: true, description: "Game Session ID")]
    #[OA\Response(response: 200, description: "Session and current question details")]
    #[OA\Response(response: 403, description: "Forbidden/Not owner of session")]
    public function show(GameSession $session)
    {
        // Add authorization check
        Gate::authorize('view', $session);

        return response()->json([
            'session' => $session,
            'question' => $this->gameService->getNextQuestion($session)
        ]);
    }

    #[OA\Post(path: "/games/{session}/answer", summary: "Submit an answer for the current question", tags: ["Game"])]
    #[OA\Parameter(name: "session", in: "path", required: true, description: "Game Session ID")]
    #[OA\RequestBody(
        required: true,
        content: new OA\JsonContent(
            required: ["answer_id"],
            properties: [new OA\Property(property: "answer_id", type: "integer", example: 123)]
        )
    )]
    #[OA\Response(response: 200, description: "Answer evaluation result (correct/wrong) and optional next question")]
    #[OA\Response(response: 400, description: "Game already ended")]
    #[OA\Response(response: 403, description: "Forbidden")]
    public function answer(Request $request, GameSession $session)
    {
        Gate::authorize('update', $session);
        if ($session->status !== 'active')
            return response()->json(['error' => 'Game ended'], 400);

        $request->validate([
            'answer_id' => 'nullable|exists:answers,id',
            'time_remaining' => 'nullable|numeric'
        ]);

        $answerId = $request->input('answer_id'); // null means timed out
        $timeRemaining = (int) $request->input('time_remaining', 0);
        $result = $this->gameService->processAnswer($session, $answerId, $timeRemaining);

        // If correct, get next question immediately? Or separate call.
        // Frontend expects next_question in response often.
        // If correct or in multiplayer competition, get next question if still active
        if ($session->status === 'active') {
            $result['next_question'] = $this->gameService->getNextQuestion($session);
        }

        return response()->json(array_merge($result, ['session' => $session]));
    }

    // Simplistic lifeline implementation
    #[OA\Post(path: "/games/{session}/lifeline", summary: "Use a lifeline for the current question", tags: ["Game"])]
    #[OA\Parameter(name: "session", in: "path", required: true, description: "Game Session ID")]
    #[OA\RequestBody(
        required: true,
        content: new OA\JsonContent(
            required: ["type"],
            properties: [
                new OA\Property(property: "type", type: "string", description: "Name of lifeline (e.g. fiftyFifty, skip, audienceVote, phoneFriend)", example: "fiftyFifty"),
                new OA\Property(property: "question_id", type: "integer", description: "ID of the question (for lifelines that need it)", example: 45)
            ]
        )
    )]
    #[OA\Response(response: 200, description: "Result of using the lifeline (e.g. hidden answers or next question)")]
    #[OA\Response(response: 403, description: "Forbidden")]
    public function lifeline(Request $request, GameSession $session)
    {
        Gate::authorize('update', $session);
        $type = $request->input('type');
        $questionId = $request->input('question_id');

        $result = $this->gameService->useLifeline($session, $type, $questionId);

        return response()->json($result);
    }
}