# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Millionaire Quiz API",
        "description": "API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "operationId": "5d4fa5d5607ed675ba93907e1fc94534",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful registration"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log in to the application",
                "operationId": "67d6d0fdd54fba132e1b6f55ca254183",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login"
                    },
                    "422": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
...
        "/forgot-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Send password reset link",
                "operationId": "f46fe87b31c0dadd30642d353886e6f1",
                "responses": {
                    "200": {
                        "description": "Password reset link sent successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/reset-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Reset password with token",
                "operationId": "f93180f97c46f79e028b188d8763b2e1",
                "responses": {
                    "200": {
                        "description": "Password has been reset successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/games/{session}": {
            "get": {
                "tags": [
                    "Game"
                ],
                "summary": "Get an active game session state",
                "operationId": "11877bd8ba37bd6c962da8a8050ac224",
                "parameters": [
                    {
                        "name": "session",
                        "in": "path",
                        "description": "Game Session ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session and current question details"
                    },
                    "403": {
                        "description": "Forbidden/Not owner of session"
                    }
                }
            }
        },
        "/games/{session}/answer": {
            "post": {
                "tags": [
                    "Game"
                ],
                "summary": "Submit an answer for the current question",
                "operationId": "e1f5263ae3c44d56a1290500b64a0cd1",
                "parameters": [
                    {
                        "name": "session",
                        "in": "path",
                        "description": "Game Session ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "answer_id"
                                ],
                                "properties": {
                                    "answer_id": {
                                        "type": "integer",
                                        "example": 123
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Answer evaluation result (correct/wrong) and optional next question"
                    },
                    "400": {
                        "description": "Game already ended"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/games/{session}/lifeline": {
            "post": {
                "tags": [
                    "Game"
                ],
                "summary": "Use a lifeline for the current question",
                "operationId": "54e842a2abc137f3f10db6147d967a1c",
                "parameters": [
                    {
                        "name": "session",
                        "in": "path",
                        "description": "Game Session ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type"
                                ],
                                "properties": {
                                    "type": {
                                        "description": "Name of lifeline (e.g. fiftyFifty, skip, audienceVote, phoneFriend)",
                                        "type": "string",
                                        "example": "fiftyFifty"
                                    },
                                    "question_id": {
                                        "description": "ID of the question (for lifelines that need it)",
                                        "type": "integer",
                                        "example": 45
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Result of using the lifeline (e.g. hidden answers or next question)"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://127.0.0.1:8001/oauth/authorize",
                        "tokenUrl": "http://127.0.0.1:8001/oauth/token",
                        "refreshUrl": "http://127.0.0.1:8001/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}